+++ apps/desktop/layer/renderer/src/atoms/debug-feature.ts
@@ -1,0 +1,7 @@
+import { createAtomHooks } from "@follow/utils/jotai"
+import { getStorageNS } from "@follow/utils/ns"
+import { atomWithStorage } from "jotai/utils"
+
+export const [, , useDebugFeatureValue, getDebugFeatureValue, ,] = createAtomHooks(
+  atomWithStorage(getStorageNS("debug-feature"), {}),
+)
