+++ apps/desktop/layer/renderer/src/modules/entry-column/Items/article-item.tsx
@@ -6,11 +6,14 @@ import { ListItem } from "~/modules/entry-column/templates/list-item-template"
 import { FeedIcon } from "~/modules/feed/feed-icon"
 import { FeedTitle } from "~/modules/feed/feed-title"
 
+import { readableContentMaxWidth } from "../styles"
 import type { EntryItemStatelessProps, UniversalItemProps } from "../types"
 
 export function ArticleItem({ entryId, entryPreview, translation }: UniversalItemProps) {
   return <ListItem entryId={entryId} entryPreview={entryPreview} translation={translation} />
 }
+
+ArticleItem.wrapperClassName = readableContentMaxWidth
 
 export function ArticleItemStateLess({ entry, feed }: EntryItemStatelessProps) {
   return (
