+++ apps/desktop/layer/renderer/src/modules/entry-column/components/DateItem.tsx
@@ -39,7 +39,7 @@ const useParseDate = (date: string) =>
     }
   }, [date])
 
-const dateItemclassName = tw`relative flex items-center text-sm lg:text-base gap-1 bg-background px-4 font-bold text-text h-7`
+const dateItemclassName = tw`relative flex items-center text-sm lg:text-base gap-1 px-4 font-bold text-text h-7`
 export const DateItem = memo(({ date, view, isSticky }: DateItemProps) => {
   if (view === FeedViewType.SocialMedia) {
     return <SocialMediaDateItem date={date} className={dateItemclassName} isSticky={isSticky} />
@@ -92,7 +92,7 @@ const DateItemInner: FC<DateItemInnerProps> = ({
   )
   return (
     <div
-      className={cn(className, isSticky && "border-b")}
+      className={cn(className, isSticky && "bg-background border-b")}
       onClick={stopPropagation}
       onMouseEnter={removeConfirm.cancel}
       onMouseLeave={removeConfirm}
