+++ apps/desktop/layer/renderer/src/modules/entry-column/styles.ts
@@ -1,1 +1,4 @@
 export const girdClassNames = tw`grid grid-cols-1 @lg:grid-cols-2 @3xl:grid-cols-3 @6xl:grid-cols-4 @7xl:grid-cols-5 gap-1.5`
+
+// Shared max-width styles for readable content
+export const readableContentMaxWidth = tw`max-w-[65ch] mx-auto px-4`
