+++ apps/desktop/layer/renderer/src/modules/entry-content/actions/header-actions.tsx
@@ -30,6 +30,7 @@ export const EntryHeaderActions = ({
           clickableDisabled={config.disabled}
           highlightMotion={config.notice}
           id={`${config.entryId}/${config.id}`}
+          size={compact ? "sm" : "base"}
         />
       )
     })
