+++ apps/desktop/layer/renderer/src/pages/(main)/(layer)/timeline/[timelineId]/[feedId]/layout.tsx
@@ -1,1 +1,5 @@
-export { CenterColumnLayout as Component } from "~/modules/app-layout/timeline-column/index"
+import { withFeature } from "~/lib/features"
+import { CenterColumnLayout } from "~/modules/app-layout/timeline-column/index"
+import { EntryColumnLayout } from "~/modules/entry-column/EntryColumnLayout"
+
+export const Component = withFeature("ai")(EntryColumnLayout, CenterColumnLayout)
