+++ packages/internal/components/package.json
@@ -24,6 +24,7 @@
   "dependencies": {
     "@essentials/request-timeout": "1.3.0",
     "@follow/hooks": "workspace:*",
+    "@follow/models": "workspace:*",
     "@follow/types": "workspace:*",
     "@follow/utils": "workspace:*",
     "@headlessui/react": "2.2.4",
